fix(scan): drop the importlib cycle-hiding hack for the L4/L5 estimator fix
service_scan_estimate.py split the per-layer cost estimator into its own
module to make line-budget room for the _TU_UNKNOWN_NOTE_SUFFIX
propagation fix, using importlib.import_module() to dodge the resulting
two-module import cycle (estimate_scan needs several service_scan
private helpers, and service_scan.run_scan() calls estimate_scan back).
That passed both the architecture gate and the AI-readiness cycle-growth
check, but only because importlib.import_module() is an ordinary
function call invisible to their static AST-based cycle detectors --
Codex correctly flagged it as hiding a genuinely new bidirectional
coupling rather than reusing the codebase's existing one-directional
importlib precedents (service.py's _service_header_scoped,
cli_buildsource.py's back-compat shim), which never have a dependency
back from their "downstream" side.

Revert the split: merge service_scan_estimate.py's content back into
service_scan.py (the L4/L5 fix logic itself is unchanged, just
relocated), and instead extract pair_wide_cxx20_std_override into a new
leaf module, cxx20_pair_dialect.py. That function has zero internal
callers within service_scan.py and no dependency on anything
service_scan-specific, so it moves with a genuine one-directional
re-export and no cycle to hide.
